--- nursery/components/CheckChild.tsx
@@ -1,28 +1,0 @@
-import { ChildrenInfo } from '@/types/children'
-import React from 'react'
-import { format } from 'date-fns';
-import { checkInChild, checkOutChild } from '@/actions/children';
-
-export default function CheckChild(child: ChildrenInfo, refreshCheckIn: any) {
-
-  const now = format(Date(), "HH:mm")
-  const [time, setTime] = React.useState<string>(now)
-
-  return (
-    <div className=''>
-      {child?.child.checkedIn ?
-        <button className="rounded-lg border-b-0 border-r-2 m-3 p-3 bg-white"
-          onClick={() => {checkOutChild(child?.child.childId)}}>Check Out Child now</button>
-        :
-        <div className='flex flex-col items-center'>
-          <button className="rounded-lg border-b-0 border-r-2 m-3 p-3 bg-white"
-            onClick={() => {checkInChild(child?.child.childId, time)}
-              
-            }>Check In Child at
-          </button>
-          <input value={time} onChange={(ev) => { setTime(ev.target.value) }} aria-label="Time" type="time" />
-        </div>
-      }
-    </div>
-  )
-}
