+++ nursery/package.json
@@ -12,7 +12,8 @@
     "date-fns": "^4.1.0",
     "next": "15.1.4",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-intersection-observer": "^9.15.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
