+++ nursery/pages/index.tsx
@@ -1,20 +1,16 @@
-import styles from "./page.module.css";
-import type { InferGetStaticPropsType } from 'next'
 import { Child, ChildrenInfo } from '../types/children'
 import ChildInfo from "@/components/ChildInfo";
 import ChildIntro from "@/components/ChildIntro";
-import CheckChild from "@/components/CheckChild";
 import 'tailwindcss/tailwind.css'
 import { useEffect, useState } from "react";
+import { format } from 'date-fns';
+import { checkInChild, checkOutChild } from '@/actions/children';
+import { useInView } from 'react-intersection-observer'
 
 export default function Home() {
 
   const [children, setChildren] = useState<Child[]>([])
-  const [refresh, setRefresh] = useState(false);
-
-  const refreshCheckIn = (value:boolean) => {
-    setRefresh(value)
-  }
+  const { ref, inView } = useInView()
 
   useEffect(() => {
     const fetchChildren = async () => {
@@ -34,30 +30,57 @@ export default function Home() {
     }
     fetchChildren()
   },
-    [refresh])
+    [])
 
-  const [displayChildren, setDisplayChildren] = useState<number>(10)
+  useEffect(() => {
+    if (inView) { setDisplayChildren(displayChildren => displayChildren + 2) }
+  }, [inView])
+
+  const [displayChildren, setDisplayChildren] = useState<number>(5)
 
   const slicedArray = children?.slice(0, displayChildren)
 
+  const now = format(Date(), "HH:mm")
+  const [time, setTime] = useState<string>(now)
+
   return (
-    <div className="bg-gray-400">
+    <div className="flex flex-col bg-gray-400">
       <h1 className="font-extrabold text-5xl text-center">Nursery</h1>
-      <div className="grid md:grid-cols-2 lg:grid-cols-2 xl:grid-cols-3 grid-rows-3 gap-4
-      p-2" >
+      <div className="grid md:grid-cols-2 lg:grid-cols-2 xl:grid-cols-3 grid-rows-3 gap-4 p-2" >
         {slicedArray?.map((child: Child) => (
           <div key={child.childId} className="flex justify-around border-2 rounded-lg  m-3 p-3">
             <ChildIntro child={child} />
             <ChildInfo child={child} />
-            <CheckChild child={child} refreshCheckIn={refreshCheckIn} />
+
+            <div className=''>
+              {child.checkedIn ?
+                <button className="rounded-lg border-b-0 border-r-2 m-3 p-3 bg-white"
+                  onClick={() => {
+                    checkOutChild(child.childId)
+                  }}>Check Out Child now at {time}</button>
+                :
+                <div className='flex flex-col items-center'>
+                  <button className="rounded-lg border-b-0 border-r-2 m-3 p-3 bg-white"
+                    onClick={() => {
+                      checkInChild(child.childId, time)
+                    }
+                    }>Check In Child at
+                  </button>
+                  <input value={time} onChange={(ev) => { setTime(ev.target.value) }} aria-label="Time" type="time" />
+                </div>
+              }
+            </div>
           </div>
         ))
         }
       </div>
       <div className="flex justify-center p-4">
-        <button onClick={() => setDisplayChildren(displayChildren => displayChildren + 5)} className="bg-blue-500 rounded-md text-3xl hover:bg-blue-700 text-white font-bold py-2 px-4">
+        <button onClick={() => setDisplayChildren(displayChildren => displayChildren + 2)} className="bg-blue-500 rounded-md text-3xl hover:bg-blue-700 text-white font-bold py-2 px-4">
           Display more children
         </button>
+      </div>
+      <div ref={ref}>
+        Loading...
       </div>
     </div>
   );
